Pull full screen modal/sheets to Root modal flow in nav stack
