+++ app/components/Views/Asset/index.js
@@ -28,6 +28,7 @@ import {
   findBlockExplorerForRpc,
   isMainnetByChainId,
 } from '../../../util/networks';
+import Routes from '../../../constants/navigation/Routes';
 
 const createStyles = (colors) =>
   StyleSheet.create({
@@ -135,8 +136,12 @@ class Asset extends PureComponent {
         colors,
         shouldShowMoreOptionsInNavBar
           ? () =>
-              navigation.navigate('AssetOptions', {
-                isNativeCurrency: isNativeToken,
+              navigation.navigate(Routes.MODAL.ROOT_MODAL_FLOW, {
+                screen: 'AssetOptions',
+                params: {
+                  isNativeCurrency: isNativeToken,
+                  address: route.params?.address,
+                },
               })
           : undefined,
         true,
